+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+* fix `PopulationComparisonProcessor`: fix measurement of **Absolute Coveredness** to not increase in case of a duplicated resource in one dataset without a corresponding resource in any other dataset  
+
 ## [3.1.1] - 2024-11-13
 
 ### Fixed